+++ client/flutter/divkit/lib/src/core/protocol/div_context.dart
@@ -1,6 +1,7 @@
 import 'package:divkit/divkit.dart';
 import 'package:divkit/src/core/protocol/div_data_provider.dart';
 import 'package:divkit/src/core/protocol/div_patch.dart';
+import 'package:divkit/src/core/protocol/div_visibility_action.dart';
 import 'package:divkit/src/utils/div_focus_node.dart';
 import 'package:flutter/widgets.dart';
 
@@ -12,6 +13,8 @@ abstract class DivContext {
   DivStateManager get stateManager;
 
   DivActionHandler get actionHandler;
+
+  DivVisibilityActionManager get visibilityActionManager;
 
   DivTimerManager get timerManager;
 
@@ -60,6 +63,15 @@ class DivRootContext extends DivContext {
 
   set patchManager(DivPatchManager value) => _patchManager = value;
 
+  DivVisibilityActionManager? _visibilityActionManager;
+
+  @override
+  DivVisibilityActionManager get visibilityActionManager =>
+      _visibilityActionManager!;
+
+  set visibilityActionManager(DivVisibilityActionManager value) =>
+      _visibilityActionManager = value;
+
   DivRootContext({
     this.buildContext,
   });
@@ -74,6 +86,7 @@ class DivRootContext extends DivContext {
 
   Future<void> dispose() async {
     _stateManager?.dispose();
+    _visibilityActionManager?.dispose();
     await _timerManager?.dispose();
     await _variableManager?.dispose();
     await dataProvider?.dispose();
@@ -84,6 +97,7 @@ class DivRootContext extends DivContext {
     _variableManager = null;
     _actionHandler = null;
     _patchManager = null;
+    _visibilityActionManager = null;
 
     // Clear the expression resolver.
     await exprResolver.clearVariables();