+++ client/flutter/divkit/lib/src/core/visibility/models/action_visibility_state_type.dart
@@ -1,0 +1,5 @@
+enum ActionVisibilityStateType {
+  end,
+  wait,
+  notStart;
+}