+++ client/flutter/divkit/lib/src/core/widgets/base/div_base_widget.dart
@@ -6,6 +6,7 @@ import 'package:divkit/src/generated_sources/div_base.dart';
 import 'package:divkit/src/utils/div_focus_node.dart';
 import 'package:divkit/src/utils/size_converters.dart';
 import 'package:flutter/widgets.dart';
+import 'package:divkit/src/core/widgets/div_visibility_emitter.dart';
 
 class DivBaseWidget extends StatefulWidget {
   final DivBase data;
@@ -70,8 +71,10 @@ class _DivBaseWidgetState extends State<DivBaseWidget> {
               height: model.height,
               width: model.width,
               alignment: model.alignment,
-              child: IgnorePointer(
-                ignoring: !model.isVisible,
+              child: DivVisibilityEmitter(
+                visibilityActions: model.visibilityActions,
+                divVisibility: model.divVisibility,
+                id: model.divId.toString(),
                 child: Opacity(
                   opacity: model.opacity,
                   child: DivTapActionEmitter(