+++ client/flutter/divkit/lib/src/core/widgets/input/div_input_model.dart
@@ -1,4 +1,4 @@
-import 'package:divkit/src/core/action/action.dart';
+import 'package:divkit/src/core/action/models/action.dart';
 import 'package:divkit/src/core/protocol/div_context.dart';
 import 'package:divkit/src/generated_sources/div_input.dart';
 import 'package:divkit/src/utils/converters.dart';